+++ src/EMRDb.h
@@ -27,8 +27,8 @@ public:
         EMRTrack        *track;
         string           filename;
         struct timespec  timestamp;
-        string           db_id;
-        vector<string>   dbs;
+        string           db_id; // holds the current db which holds the track data.
+        vector<string>   dbs; // holds the ids of all the other dbs in which the track appears. 
         bool             overridden;
 
         TrackInfo(EMRTrack *_track, const string &_filename, const struct timespec &_timestamp, string db_id) :
@@ -40,7 +40,7 @@ public:
     static const string LOGICAL_TRACK_FILE_EXT;
 
     ~EMRDb();
-    //Only called by NRTest - here for compilation, should be removed
+    
     const string &grootdir() { return m_rootdirs[0]; }
     const string &urootdir() { return m_rootdirs[0]; }
     const char* dob_trackname() { return DOB_TRACKNAME; }
@@ -88,7 +88,7 @@ public:
     // If track list file is missing => builds it.
     // Removes outdated tracks from the memory.
 
-    void init(vector<string> rootdirs, vector<bool> dirs_load_on_demand, bool do_reload);
+    void init(const vector<string>& rootdirs, const vector<bool>& dirs_load_on_demand, const bool& do_reload);
 
     // Same as init, but roots must be set already.
     // Intended to be called at the beginning of each transaction.
@@ -104,17 +104,17 @@ public:
     // with the tracks.
     void reload();
 
-    void load_track(const char *track_name, string db_id);
+    void load_track(const char *track_name, const string& db_id);
     
     // Unloads the track from internal state and updates .naryn file
     // If soft is set to true, unload track does not update .naryn file, used mainly for overriding mechanism
-    void unload_track(const char *track_name, bool overridden=true, bool soft=false);
+    void unload_track(const char *track_name, const bool& overridden=true, const bool& soft=false);
 
     // Add a logical track to the database
-    void add_logical_track(const char *track_name, const char *source_name, const vector<int> &values, bool create_file, bool update);
-    void add_logical_track(const char *track_name, const char *source_name, bool create_file, bool update);
+    void add_logical_track(const char *track_name, const char *source_name, const vector<int> &values, const bool& create_file, const bool& update);
+    void add_logical_track(const char *track_name, const char *source_name, const bool& create_file, const bool& update);
 
-    void remove_logical_track(const char *track_name, bool update);
+    void remove_logical_track(const char *track_name, const bool& update);
 
     // Writes the logical track list file. If the file doesn't exist - recreates it. 
     void update_logical_tracks_file();
@@ -147,7 +147,7 @@ public:
     void ids_subset(vector<unsigned> &ids, const char *src, double fraction, bool complementary);
     bool is_in_subset(unsigned id) const { return m_ids_subset.empty() || m_ids_subset.find(id) != m_ids_subset.end(); }
     void clear_ids_subset(bool warn);
-    int get_db_idx(string db_id);
+    int get_db_idx(const string& db_id);
 
 protected:
 	typedef unordered_map<string, TrackInfo> Name2Track;
@@ -190,6 +190,7 @@ protected:
     
     string track_filename(string db_id, const string &track_name) const { return db_id + string("/") + track_name + TRACK_FILE_EXT; }
     string track_attrs_filename(string db_id, const string &track_name) const { return db_id + string("/.") + track_name + TRACK_ATTRS_FILE_EXT; }
+    string logical_track_attrs_filename(const string &track_name) const { return logical_tracks_dir() + string("/.") + track_name + TRACK_ATTRS_FILE_EXT; }
     string logical_tracks_dir() const { return m_rootdirs[0] + string("/logical"); }    
     string track_list_filename(string db_id) const { return db_id + "/" + TRACK_LIST_FILENAME; }
     string tracks_attrs_filename(string db_id) const { return db_id + "/" + TRACKS_ATTRS_FILENAME; }