+++ tests/testthat/helper-logical_tracks.R
@@ -25,3 +25,19 @@ logical_track_path <- function(track) {
     root <- get("EMR_GROOT", envir = .GlobalEnv)
     file.path(root, "logical", glue::glue("{track}.ltrack"))
 }
+
+logical_track_ok <- function(track, source, values = NULL) {
+    expect_true(track %in% emr_track.logical.ls())
+    expect_true(track %in% emr_track.ls())
+    expect_true(track %in% emr_track.global.ls())
+    expect_true(emr_track.logical.exists(track))
+    expect_equal(emr_track.logical.info(track)$source, source)
+    if (is.null(values) || length(values) == 0) {
+        expect_null(emr_track.logical.info(track)$values)
+    } else {
+        expect_equal(emr_track.logical.info(track)$values, values)
+    }
+
+    expect_true(emr_track.exists(track))
+    expect_true(file.exists(logical_track_path(track)))
+}