+++ tests/testthat/test-summary.R
@@ -1,3 +1,5 @@
+load_test_db()
+
 
 test_that("emr_summary works", {
     expect_equal(