+++ tests/testthat/test-vtrack.R
@@ -1,3 +1,5 @@
+load_test_db()
+
 
 test_that("emr_vtrack fails when track deosn't exist", {
     expect_error(emr_vtrack.create("v1", "blabla"))